Implement new test case for dashboard chunk size
